+++ pilot_drive/utils/adb_read_icon_db.py
