+++ pilot_drive/utils/bt_ctl.py
@@ -1,6 +1,6 @@
 # TODO: Should the function just set 
 
-from utils.bt_track import Track
+from pilot_drive.utils.bt_track import Track
 from dbus.mainloop.glib import DBusGMainLoop    # Handling the DBus event based loop
 from gi.repository import GLib                  # Handling the DBus event based loop
 from time import sleep
